requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-decouple==3.8
dj-database-url==2.1.0